requirements.txt
# Phase 1 & 2 Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pdfplumber==0.10.3
pdf2image==1.16.3
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78
ultralytics==8.0.196
pytest==7.4.3
pytest-asyncio==0.21.1

# Phase 3: OCR Dependencies for Windows
paddlepaddle==2.6.2  # Latest compatible version for Windows
paddleocr==2.7.0.3  # Use specific version
pyclipper==1.3.0.post5
shapely==2.0.2
lxml==5.2.1